requirements.txt
requests
six
pyyaml
robobrowser
docopt
